+++ src/MeasureMap/ProfilerResultCollectionExtensions.cs
@@ -1,4 +1,5 @@
-﻿using System.Linq;
+﻿using MeasureMap.Tracers;
+using System.Linq;
 using System.Text;
 
 namespace MeasureMap
@@ -11,32 +12,52 @@ namespace MeasureMap
         /// <summary>
         /// Traces the output of a Benchmark Test
         /// </summary>
-        /// <param name="results"></param>
+        /// <param name="result"></param>
         /// <returns></returns>
-        public static string Trace(this IProfilerResultCollection results)
+        public static void Trace(this IProfilerResultCollection result)
         {
-            var sb = new StringBuilder();
-
-            sb.AppendLine("## MeasureMap Benchmark");
-            sb.AppendLine($" Iterations:\t\t{results.Iterations}");
-            sb.AppendLine($"### Summary");
-            sb.AppendLine("| Name | Avg Time | Avg ms | Avg Ticks | Total | Fastest | Slowest | Iterations | Throughput |");
-            sb.AppendLine("|--- |---: |---: |---: |---: |---: |---: |---: |---: |");
-            foreach (var key in results.Keys)
-            {
-                sb.AppendLine($"| {key} {TraceLine(results[key])}");
-            }
-
-            var result = sb.ToString();
-
-            System.Diagnostics.Trace.WriteLine(result);
-
-            return result;
+            result.Trace(TraceOptions.Default.Tracer, TraceOptions.Default.ResultWriter);
         }
 
-        private static string TraceLine(IProfilerResult result)
+        /// <summary>
+        /// Trace the output of a Benchmark Test
+        /// </summary>
+        /// <param name="result"></param>
+        /// <param name="options"></param>
+        public static void Trace(this IProfilerResultCollection result, TraceOptions options)
         {
-            return $"| {result.AverageTime} | {result.AverageMilliseconds} | {result.AverageTicks} | {result.TotalTime.ToString()} | {result.Fastest.Ticks} | {result.Slowest.Ticks} | {result.Iterations.Count()} | {result.Throughput()} |";
+            result.Trace(options.Tracer, options.ResultWriter);
+        }
+
+        /// <summary>
+        /// Trace the output of a Benchmark Test
+        /// </summary>
+        /// <param name="result"></param>
+        /// <param name="tracer"></param>
+        public static void Trace(this IProfilerResultCollection result, ITracer tracer)
+        {
+            result.Trace(tracer, TraceOptions.Default.ResultWriter);
+        }
+
+        /// <summary>
+        /// Trace the output of a Benchmark Test
+        /// </summary>
+        /// <param name="result"></param>
+        /// <param name="writer"></param>
+        public static void Trace(this IProfilerResultCollection result, IResultWriter writer)
+        {
+            result.Trace(TraceOptions.Default.Tracer, writer);
+        }
+
+        /// <summary>
+        /// Trace the output of a Benchmark Test
+        /// </summary>
+        /// <param name="result"></param>
+        /// <param name="tracer"></param>
+        /// <param name="writer"></param>
+        public static void Trace(this IProfilerResultCollection result, ITracer tracer, IResultWriter writer)
+        {
+            tracer.Trace(result, writer);
         }
     }
 }