+++ src/MeasureMap/Tracers/ITracer.cs
@@ -1,0 +1,23 @@
+﻿
+namespace MeasureMap.Tracers
+{
+    /// <summary>
+    /// Tracers are used to write the result in a desired format to a output source
+    /// </summary>
+    public interface ITracer
+    {
+        /// <summary>
+        /// Trace the <see cref="IProfilerResult"/>
+        /// </summary>
+        /// <param name="result"></param>
+        /// <param name="writer"></param>
+        void Trace(IProfilerResult result, IResultWriter writer);
+
+        /// <summary>
+        /// Trace the <see cref="IProfilerResultCollection"/>
+        /// </summary>
+        /// <param name="result"></param>
+        /// <param name="writer"></param>
+        void Trace(IProfilerResultCollection result, IResultWriter writer);
+    }
+}